🩹 fix 유저 공격 변경
